+++ lib/scss_lint/linter/property_units.rb
@@ -4,13 +4,17 @@ module SCSSLint
     include LinterRegistry
 
     NUMBER_WITH_UNITS_REGEX = /
-      (?:^|\s)    # beginning of value or whitespace
       (?:
-        \d+       # any number of digits, e.g. 123
-        |         # or
-        \d*\.?\d+ # any number of digits with decimal, e.g. 1.23 or .123
+       (["']).+?\1 # [0: quote mark] quoted string, e.g. "hi there"
+       |           # or
+       (?:^|\s)    # beginning of value or whitespace
+       (?:
+        \d+        # any number of digits, e.g. 123
+        |          # or
+        \d*\.?\d+  # any number of digits with decimal, e.g. 1.23 or .123
+       )
+       ([a-z%]+)   # [1: units] letters or percent sign, e.g. px or %
       )
-      ([a-z%]+)   # letters or percent sign, e.g. px or %
     /ix
 
     def visit_root(_node)
@@ -29,8 +33,11 @@ module SCSSLint
       end
 
       if node.value.respond_to?(:value)
-        node.value.value.to_s.scan(NUMBER_WITH_UNITS_REGEX).each do |units|
-          check_units(node, property, units[0])
+        node.value.value.to_s.scan(NUMBER_WITH_UNITS_REGEX).each do |matches|
+          is_quoted_value = !matches[0].nil?
+          next if is_quoted_value
+          units = matches[1]
+          check_units(node, property, units)
         end
       end