+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 * Add `TrailingWhitespace` linter which reports whitespace characters at the
   end of a line
 * Change `NestingDepth` to allow parent selectors to be ignored in depth count
-  by passing configuration option `ignore_parent_selectors: true`
+  via the `ignore_parent_selectors` configuration option
 
 ## 0.39.0