+++ CHANGELOG.md
@@ -17,6 +17,8 @@
   end of a line
 * Change `NestingDepth` to allow parent selectors to be ignored in depth count
   via the `ignore_parent_selectors` configuration option
+* Fix `PropertyUnits` for properties that have multiple unit-like values (e.g.
+  shorthand properties)
 
 ## 0.39.0